Speech_Features_Extraction/Acoustic_Features/requirements.txt
numpy>=1.23
pandas>=1.5
# audio I/O + conversion
pydub>=0.25
soundfile>=0.12
ffmpeg-python>=0.2
# acoustic features (pyopenSMILE wrapper)
opensmile>=2.5
tqdm>=4.66
pyyaml>=6.0
joblib>=1.3